Various bug fixes (#957)
* Fixed a million small depths description bugs
* Fixed lilypads being unplaceable/unbreakable on plots
* fixed death sort caring about compasses
* Fixed region scaling with hunting companion
* Fixed quickdraw with shattered items

+++ plugins/paper/src/main/java/com/playmonumenta/plugins/depths/abilities/dawnbringer/SoothingCombos.java
@@ -68,7 +68,7 @@ public class SoothingCombos extends DepthsAbility {
 
 	@Override
 	public String getDescription(int rarity) {
-		return "Every third melee attack applies " + DepthsUtils.getRarityColor(rarity) + DepthsUtils.roundPercent(SPEED_PERCENT[rarity - 1]) + "%" + ChatColor.WHITE + " speed and Haste I for " + DepthsUtils.getRarityColor(rarity) + DURATION[rarity - 1] + ChatColor.WHITE + " seconds to players within " + RANGE + " blocks, including the user.";
+		return "Every third melee attack applies " + DepthsUtils.getRarityColor(rarity) + DepthsUtils.roundPercent(SPEED_PERCENT[rarity - 1]) + "%" + ChatColor.WHITE + " speed and Haste I for " + DepthsUtils.getRarityColor(rarity) + (int) DURATION[rarity - 1] + ChatColor.WHITE + " seconds to players within " + RANGE + " blocks, including the user.";
 	}
 
 	@Override
