+++ plugins/paper/src/main/java/com/playmonumenta/plugins/depths/abilities/flamecaller/RingOfFlames.java
@@ -36,7 +36,7 @@ public class RingOfFlames extends DepthsAbility {
 
 	public static final String ABILITY_NAME = "Ring of Flames";
 	private static final int COOLDOWN = 18 * 20;
-	private static final double[] DAMAGE = {4, 5, 6, 7, 8, 10};
+	private static final int[] DAMAGE = {4, 5, 6, 7, 8, 10};
 	private static final int[] DURATION = {8 * 20, 9 * 20, 10 * 20, 11 * 20, 12 * 20, 14 * 20};
 	private static final int EFFECT_DURATION = 4 * 20;
