+++ plugins/paper/src/main/java/com/playmonumenta/plugins/depths/abilities/flamecaller/VolcanicCombos.java
@@ -22,8 +22,8 @@ import net.md_5.bungee.api.ChatColor;
 
 public class VolcanicCombos extends DepthsAbility {
 	public static final String ABILITY_NAME = "Volcanic Combos";
-	public static final double[] DAMAGE = {6, 7, 8, 9, 10, 12};
-	public static final double RADIUS = 4;
+	public static final int[] DAMAGE = {6, 7, 8, 9, 10, 12};
+	public static final int RADIUS = 4;
 	public static final int FIRE_TICKS = 3 * 20;
 
 	private int mComboCount = 0;
